Bug Fix and Restructure
Fixed ext being appended to file names per instance run
moved header files into separate folder
