+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(FileGenny C)
 
 set(CMAKE_C_STANDARD 11)
 
-add_executable(FileGenny main.c GenerateDummyText.c GenerateDummyText.h FileFactory.c FileFactory.h SupportedFileTypes.h FileQueue.c FileQueue.h)
+add_executable(FileGenny main.c GenerateDummyText.c HeaderFiles/GenerateDummyText.h FileFactory.c HeaderFiles/FileFactory.h HeaderFiles/SupportedFileTypes.h FileQueue.c HeaderFiles/FileQueue.h)
