+++ GenerateDummyText.c
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 #include <time.h>
 #include <stdbool.h>
-#include "GenerateDummyText.h"
+#include "HeaderFiles/GenerateDummyText.h"
