+++ HeaderFiles/SupportedFileTypes.h
