+++ cmake-build-debug/CMakeFiles/FileGenny.dir/C.includecache
@@ -11,38 +11,18 @@ stdbool.h
 -
 stdio.h
 -
-SupportedFileTypes.h
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/SupportedFileTypes.h
-FileFactory.h
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.h
+HeaderFiles/SupportedFileTypes.h
+/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/SupportedFileTypes.h
+HeaderFiles/FileFactory.h
+/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileFactory.h
 string.h
 -
 stdlib.h
 -
 
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.h
+/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileFactory.h
 stdbool.h
 -
 
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileQueue.h
-
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/GenerateDummyText.h
-stdbool.h
--
-
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/SupportedFileTypes.h
-
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/main.c
-stdio.h
--
-malloc.h
--
-SupportedFileTypes.h
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/SupportedFileTypes.h
-FileFactory.h
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.h
-GenerateDummyText.h
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/GenerateDummyText.h
-FileQueue.h
-/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileQueue.h
+/cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/SupportedFileTypes.h
