+++ cmake-build-debug/CMakeFiles/FileGenny.dir/depend.internal
@@ -3,20 +3,20 @@
 
 CMakeFiles/FileGenny.dir/FileFactory.c.o
  /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.c
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.h
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/SupportedFileTypes.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileFactory.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/SupportedFileTypes.h
 CMakeFiles/FileGenny.dir/FileQueue.c.o
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.h
  /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileQueue.c
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileQueue.h
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/GenerateDummyText.h
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/SupportedFileTypes.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileFactory.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileQueue.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/GenerateDummyText.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/SupportedFileTypes.h
 CMakeFiles/FileGenny.dir/GenerateDummyText.c.o
  /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/GenerateDummyText.c
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/GenerateDummyText.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/GenerateDummyText.h
 CMakeFiles/FileGenny.dir/main.c.o
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileFactory.h
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/FileQueue.h
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/GenerateDummyText.h
- /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/SupportedFileTypes.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileFactory.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/FileQueue.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/GenerateDummyText.h
+ /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/HeaderFiles/SupportedFileTypes.h
  /cygdrive/c/Users/Marek/Desktop/IHateGit/FileGenny/main.c
