+++ cmake-build-debug/FileGenny.exe.stackdump
@@ -1,21 +1,17 @@
-Exception: STATUS_ACCESS_VIOLATION at rip=0018019A24C
-rax=0000000000000000 rbx=00000000000003E7 rcx=00000000FFFFD680
-rdx=00000000FFFFC838 rsi=000000018021AEE0 rdi=00000000FFFFC838
-r8 =00000000000003E7 r9 =0000000000000001 r10=00000000FFFFC6A0
-r11=00000000FFFFC860 r12=000000018019A230 r13=00000000FFFFD680
-r14=00000000FFFFC860 r15=00000000FFFFC860
-rbp=00000000FFFFC940 rsp=00000000FFFFC6F0
-program=C:\Users\Marek\Desktop\IHateGit\FileGenny\cmake-build-debug\FileGenny.exe, pid 547, thread main
-cs=0033 ds=002B es=002B fs=0053 gs=002B ss=002B
 Stack trace:
 Frame        Function    Args
-000FFFFC940  0018019A24C (000FFFFCDF0, 00000000000, 000000000E9, 000FFFFC860)
-000FFFFC940  001801C55EF (0018021AD70, 000000003E7, 000FFFFCAE8, 000FFFFD680)
-000FFFFCB60  0018019D92E (FFFFFFFFFFFFE22E, 006000484E0, 000000003E7, 000FFFFD680)
-000FFFFCB60  00180128DEB (FFFFFFFFFFFFE22E, 006000484E0, 000000003E7, 000FFFFD680)
-000FFFFCB60  001004012DB (000FFFFCBF0, 006000484E0, 000000003E7, 006000484E0)
-000FFFFCBC0  001004018C6 (000000003E7, 000FFFFCBF0, 000FFFFCC70, 73726573555C3A43)
-000FFFFCC20  00100401102 (00000000020, FF0700010302FF00, 0018004A7AA, 001800497F0)
+000FFFFC460  00180060BF9 (000FFFFC678, 00000000002, 00180128E35, 000FFFFDE50)
+00000000000  00180062AD7 (00000000064, 00000000000, 0000000010C, 00000000000)
+000FFFFC750  001801314DB (00000000000, 00000000000, 00000000000, 00000000000)
+000FFFFCB60  001801575E4 (000FFFFD680, 00000000000, 0000000010C, 0000000010C)
+000FFFFCB60  0018012DB7B (00000000000, 00000000000, 0060004A438, 00000000000)
+000FFFFCB60  0018012DEAF (0060004A438, 00000000000, 0060004A438, 00000000000)
+000FFFFCB60  00180169377 (00100000001, 000FFFFCC70, 0018027FB40, 000FFFFCB75)
+000FFFFCB60  001800C88DE (0018015BE8F, 0000000000A, 001801007F0, 000FFFFD680)
+000FFFFCB60  00180128DEB (0018015BE8F, 0000000000A, 001801007F0, 000FFFFD680)
+000FFFFCB60  0010040147D (000FFFFCBF0, 006000484E0, 000000003E7, 000FFFFCC70)
+000FFFFCBC0  00100401881 (000000003E7, 000FFFFCBF0, 000FFFFCC70, 73726573555C3A43)
+000FFFFCC20  001004010FB (00000000020, FF0700010302FF00, 0018004A7AA, 001800497F0)
 000FFFFCCD0  0018004A816 (00000000000, 00000000000, 00000000000, 00000000000)
 00000000000  00180048353 (00000000000, 00000000000, 00000000000, 00000000000)
 000FFFFFFF0  00180048404 (00000000000, 00000000000, 00000000000, 00000000000)
