+++ main.c
@@ -1,18 +1,16 @@
 #include <stdio.h>
 #include <malloc.h>
-#include "SupportedFileTypes.h"
-#include "FileFactory.h"
-#include "GenerateDummyText.h"
-#include "FileQueue.h"
+#include "HeaderFiles/SupportedFileTypes.h"
+#include "HeaderFiles/FileFactory.h"
+#include "HeaderFiles/GenerateDummyText.h"
+#include "HeaderFiles/FileQueue.h"
 int main() {
     int arrsize = 999;
-    char filepath[] = "C:\\Users\\Marek\\Desktop\\SampleFile\\newFile";
+    const char filepath[] = "C:\\Users\\Marek\\Desktop\\Sample\\newFile";
     printf("Begin\n");
     AddFile(txt,arrsize,filepath);
     AddFile(csv,arrsize,filepath);
-    printf("Second\n");
     AddFile(log,arrsize,filepath);
-
     AddFile(generic,arrsize,filepath);
     AddFile(json,arrsize,filepath);
     return 0;
